+++ src/bot/bot.py
@@ -1,7 +1,9 @@
 # bot.py
+from dotenv import load_dotenv
+load_dotenv()
+
 import discord
 from discord.ext import commands
-from dotenv import load_dotenv
 import os
 import logging
 
@@ -10,12 +12,11 @@ logging.basicConfig(
     format="%(asctime)s - %(levelname)s - %(message)s"
 )
 
-from db import criar_tabelas, registrar_usuario, buscar_usuario, registrar_voto, adicionar_filme, buscar_filmes_por_usuario, buscar_todos_os_filmes, buscar_todos_os_usuarios, contar_votos_recebidos_todos_usuario, contar_todos_os_votos_por_usuario, contar_generos_mais_assistidos, contar_generos_da_hora, contar_generos_lixo, contar_generos_por_usuario, buscar_filme_por_id
-from tmdb import buscar_detalhes_filme
-from sheets import adicionar_filme_na_planilha, escrever_voto_na_planilha
-from sincronizar_filmes import sincronizar_planilha
+from src.bot.db.db import criar_tabelas, registrar_usuario, buscar_usuario, registrar_voto, adicionar_filme, buscar_filmes_por_usuario, buscar_todos_os_filmes, buscar_todos_os_usuarios, contar_votos_recebidos_todos_usuario, contar_todos_os_votos_por_usuario, contar_generos_mais_assistidos, contar_generos_da_hora, contar_generos_lixo, contar_generos_por_usuario, buscar_filme_por_id
+from src.bot.tmdb import buscar_detalhes_filme
+from src.bot.sheets.sheets import adicionar_filme_na_planilha, escrever_voto_na_planilha
+from src.bot.sincronizar_filmes import sincronizar_planilha
 
-load_dotenv()
 TOKEN = os.getenv("DISCORD_TOKEN")
 
 intents = discord.Intents.default()
